Split API doc linking into explicit step
